\name{run_featuremsea_app}
\alias{run_featuremsea_app}
\title{Run featureMSEA Shiny Application}
\usage{
run_featuremsea_app(...)
}
\arguments{
\item{...}{Additional arguments passed to shiny::shinyApp}
}
\value{
A Shiny app object
}
\description{
This function launches the featureMSEA Shiny application for interactive
metabolite set enrichment analysis.
}
\examples{
\dontrun{
run_featuremsea_app()
}
}